cpu: Change GUPSGen::memSize to uint64_t
This change changes the type for GUPSGen::memSize to uint64_t.
Previously uint32_t was used that would overflow for memories
bigger than 4GB.

Reviewed-on: https://gem5-review.googlesource.com/c/public/gem5/+/52903
Maintainer: Jason Lowe-Power <power.jg@gmail.com>
Tested-by: kokoro <noreply+kokoro@google.com>

+++ src/cpu/testers/traffic_gen/gups_gen.hh
@@ -256,7 +256,7 @@ class GUPSGen : public ClockedObject
     /**
      * @brief Size of the memory in bytes that will be allocated for the array.
      */
-    const uint32_t memSize;
+    const uint64_t memSize;
 
     /**
      * @brief The number of updates to do before creating an exit event.
